requirements.txt
requests==2.32.5
PySide6==6.5.3
shiboken6==6.5.3
numpy>=1.26,<2.0
opencv-python-headless<5
Pillow>=10.2,<11
piexif>=1.1.3
pillow-heif>=0.16   # optional; enables HEIC/HEIF
